changed stuff based on comments
