+++ apps/diary/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+
+from .models import Post
+
+admin.site.register(Post)
+
