+++ apps/diary/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class DiaryConfig(AppConfig):
+    name = 'diary'
+
