+++ apps/diary/templates/diary/archive.html
@@ -1,0 +1,33 @@
+{% extends 'diary/base.html' %}
+
+{% block content %}
+
+    {% if user.is_authenticated %}
+        <div class="row">
+            <div class="card card-signin my-5">
+                <div class="card-body">
+                        <h5 class="card-title text-center">Archive of your diary entries.</h5>
+                        {% for item in object_list %}
+                            <article class="media content-section">
+                                <div class="media-body">
+                                <div class="article-metadata">
+                                    <small class="text-muted">{{ item.pub_date|date:"F d, Y" }}</small>
+                                </div>
+                                <h2><a class="article-title" href="{% url 'entry' item.id %}">{{ item.title }}</a></h2>
+                                <p class="article-content">{{ item.content }}</p>
+                                </div>
+                            </article>
+                        {% endfor %}
+                </div>
+            </div>
+        </div>
+        
+
+    {% else %}
+        You are not logged in.
+
+    {% endif %}
+
+
+
+{% endblock content %}
