+++ apps/diary/tests.py
