+++ apps/users/admin.py
@@ -1,0 +1,14 @@
+from django.contrib import admin
+from django.contrib.auth.admin import UserAdmin
+
+from .forms import CustomUserCreationForm, CustomUserChangeForm
+from .models import CustomUser
+
+class CustomUserAdmin(UserAdmin):
+    add_form = CustomUserCreationForm
+    form = CustomUserChangeForm
+    list_display = ['username', 'first_name', 'last_name', 'memory', 'country']
+    model = CustomUser
+
+admin.site.register(CustomUser, CustomUserAdmin)
+
