+++ apps/users/templates/users/profile.html
@@ -1,0 +1,44 @@
+{% extends 'users/base.html' %}
+
+{% block content %}
+
+    {% if user.is_authenticated %}
+        <div class="row">
+            <div class="col-md-4"></div>
+            <div class="col-md-4">
+                <div class="card card-signin my-5">
+                    <div class="card-body">
+                        <h3 class="card-title text-center">{{object.first_name}} {{ object.last_name }}</h5>
+                        
+                        <h4> {{ object.gender}} </h4> &nbsp; <h4> {{ object.age }} </h4> <br>
+                        <h4> Country: <em> {{ object.country }} </em></h4> <h5>Memory: <em> {{ object.memory }} </em></en></h5>
+                        <br><br>
+                        <p> {{ object.about }} </p>
+
+
+                        
+                    </div>
+                </div>
+            </div>
+            <div class="col-md-4"></div>
+        </div>
+    {% else %}
+        <div class="row">
+            <div class="col-md-3"></div>
+            <div class="col-md-6">
+                <div class="card card-signin my-5">
+                    <div class="card-body">
+                        <h2 class="font-weight-bold">Hold it right there!</h2>
+                        <h4 class="text-danger">You are not allowed on this page.</h4>
+                        <p class="text-muted">Please click the link below to log in.</p>
+                        <br>
+                        <a class="btn btn-primary btn-md text-center align-center" href="{% url 'login' %}" role="button">Login</a>        
+                    </div>
+                </div>
+               </div>
+
+            <div class="col-md-3"></div>
+        </div>
+    {% endif %}
+{% endblock content %}
+
