+++ apps/users/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+from . import views
+
+urlpatterns = [
+    path('signup/', views.SignUp.as_view(), name='signup'),
+    path('profile/<int:pk>/', views.ProfileView.as_view(), name='profile'),
+]
+
