+++ apps/users/views.py
@@ -1,0 +1,23 @@
+from django.shortcuts import render
+
+from django.urls import reverse_lazy
+from django.views.generic.edit import CreateView
+from django.views.generic import DetailView
+from .models import CustomUser
+
+from .forms import CustomUserCreationForm
+# Create your views here.
+
+class SignUp(CreateView):
+    form_class = CustomUserCreationForm
+    success_url = reverse_lazy('login')
+    template_name = 'users/signup.html'
+    
+
+class ProfileView(DetailView):
+    template_name = 'users/profile.html'
+    model = CustomUser
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context['user'] = self.request.user
+        return context
