--- diary/admin.py
@@ -1,8 +1,0 @@
-from django.contrib import admin
-
-from .models import Post, Tag
-# Register your models here.
-
-admin.site.register(Post)
-admin.site.register(Tag)
-
