--- diary_proj/urls.py
@@ -1,11 +1,0 @@
-
-from django.contrib import admin
-from django.urls import path,include
-
-
-urlpatterns = [
-    path('', include('diary.urls')),
-    path('admin/', admin.site.urls),
-    path('users/', include('users.urls')),
-    path('users/', include('django.contrib.auth.urls')),
-]
