--- templates/add_entry.html
@@ -1,27 +1,0 @@
-{% extends 'base.html' %}
-
-{% load crispy_forms_tags %}
-{% block content %} 
-    {% if user.is_authenticated %}
-        <div class="row">
-            <div class="col-md-2"></div>
-            <div class="col-md-8">
-                <div class="card card-signin my-5">
-                    <div class="card-body">
-                        <h5 class="card-title text-center">Add an Entry</h5>
-                        <form class="form-signin" method="POST">
-                            {% csrf_token %}
-                            {{ form|crispy }}
-                            <input type="submit" value="Add Post!" class="btn btn-lg btn-primary btn-block text-uppercase"/>
-                        </form>
-    
-                    </div>
-                </div>
-            </div>
-            <div class="col-md-2"></div>
-        </div>
-    {% else %}
-
-    {% endif %}
-
-{% endblock content %}
