--- templates/base.html
@@ -1,67 +1,0 @@
-{% load static %}
-
-<!DOCTYPE html> 
-<html>
-    <head>
-        <meta charset="utf-8">
-        <title>Diary App!</title>
-        <link rel="stylesheet" href="{% static 'css/base.css' %}">
-        <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
-
-    </head>
-    <body>
-            <nav class="navbar navbar-expand-lg navbar-light bg-light pushdown">
-                    <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarTogglerDemo01" aria-controls="navbarTogglerDemo01" aria-expanded="false" aria-label="Toggle navigation">
-                      <span class="navbar-toggler-icon"></span>
-                    </button>
-                    <div class="collapse navbar-collapse" id="navbarTogglerDemo01">
-                      <a class="navbar-brand" href="#">Diary App!</a>
-                      <ul class="navbar-nav mr-auto mt-2 mt-lg-0">
-                        <li class="nav-item">
-                            <a class="nav-link" href="{% url 'home' %}">Home</a>
-                        </li>
-                        {% if user.is_authenticated %}
-                        <li class="nav-item">
-                          <a class="nav-link" href="{% url 'addentry' %}">Add Entry</a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link" href="{% url 'archive' %}">Archive</a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link" href="{% url 'profile' %}">My Profile</a>
-                        </li>
-                        {% else %}
-                        <li class="nav-item">
-                            <a class="nav-link" href="#">Add Entry</a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link" href="#">Archive</a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link" href="#">My Profile</a>
-                        </li>
-                        <li class="nav-item">
-                          <a class="nav-link disabled" href="#">Disabled</a>
-                        </li>
-                        {% endif %}
-                      </ul>
-                    
-                      <form class="form-inline ml-auto">
-                            {% if user.is_authenticated %}
-                            <a href="{% url 'logout' %}" class="btn btn-outline-secondary">Log out</a>
-                            {% else %}
-                            <a href="{% url 'login' %}" class="btn btn-primary">Log in</a>
-                            {% endif %}
-                        </form>
-
-                    </div>
-            </nav>
-
-
-
-        <div class="container">
-            {% block content %}
-            {% endblock content %}
-        </div>
-    </body>
-</html>
