--- templates/detail_entry.html
@@ -1,69 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-    
-    {% if user.is_authenticated %}
-        {% if object %}
-            <div class="row">
-                <div class="col-md-4"></div>
-                <div class="col-md-4">
-                    <div class="card card-signin my-5">
-                        <div class="card-body">
-                            <h3 class="card-title text-center">{{object.title}}</h5>
-                            
-                            <h5>by {{ object.author }}</h5>
-                            <h5>written on: {{ object.pub_date }} </h5>
-                            <p>{{ object.content }}</p>
-                            <br><br>
-                            <h4>tags:</h4>
-                            <ul>
-                            {% for item in tags %}
-                                <li>{{item}}</li>
-                            {% endfor %}
-                            </ul>
-                        </div>
-                    </div>
-                </div>
-                <div class="col-md-4"></div>
-            </div>
-        {% else %}
-            <div class="row">
-                <div class="col-md-3"></div>
-                <div class="col-md-6">
-                    <div class="card card-signin my-5">
-                        <div class="card-body">
-                            <h2 class="font-weight-bold">Hold it right there!</h2>
-                            <h4 class="text-danger">This is someone else's diary entry.</h4>
-                        </div>
-                    </div>
-                </div>
-                <div class="col-md-3"></div>
-            </div>
-        {% endif %}
-
-    {% else %}
-        Oops! You are not logged in.
-
-    {% endif %}
-
-{% endblock content %}
-
-
-<div class="row">
-        <div class="card card-signin my-5">
-            <div class="card-body">
-                    <h5 class="card-title text-center">Archive of your diary entries.</h5>
-                    {% for item in object_list %}
-                        <article class="media content-section">
-                            <div class="media-body">
-                            <div class="article-metadata">
-                                <small class="text-muted">{{ item.pub_date|date:"F d, Y" }}</small>
-                            </div>
-                            <h2><a class="article-title" href="{% url 'entry' item.id %}">{{ item.title }}</a></h2>
-                            <p class="article-content">{{ item.content }}</p>
-                            </div>
-                        </article>
-                    {% endfor %}
-            </div>
-        </div>
-    </div>
