--- templates/profile.html
@@ -1,30 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-
-    {% if user.is_authenticated %}
-        <h2>Profile</h2><br><br>
-        <h3>{{ user.first_name }} {{ user.last_name }}</h3><br><br>
-        <h3> {{ user.gender }}, {{ user.age }}, {{ user.country }}  </h3> <br><br>
-        <p> {{ user.about }} </p>
-        <h4> {{ user.memory }} </h4>
-
-    {% else %}
-        <div class="row">
-            <div class="col-md-3"></div>
-            <div class="col-md-6">
-                <div class="card card-signin my-5">
-                    <div class="card-body">
-                        <h2 class="font-weight-bold">Hold it right there!</h2>
-                        <h4 class="text-danger">You are not allowed on this page.</h4>
-                        <p class="text-muted">Please click the link below to log in.</p>
-                        <br>
-                        <a class="btn btn-primary btn-md text-center align-center" href="{% url 'login' %}" role="button">Login</a>        
-                    </div>
-                </div>
-               </div>
-
-            <div class="col-md-3"></div>
-        </div>
-    {% endif %}
-{% endblock content %}
