--- templates/signup.html
@@ -1,24 +1,0 @@
-{% extends 'base.html' %}
-
-{% load crispy_forms_tags %}
-{% block title %}Sign Up {% endblock %}
-
-{% block content %}
-    <div class="row">
-        <div class="col-md-2"></div>
-        <div class="col-md-8">
-            <div class="card card-signin my-5">
-                <div class="card-body">
-                    <h5 class="card-title text-center">Sign up</h5>
-                    <form class="form-signin" method="POST">
-                        {% csrf_token %}
-                        {{ form|crispy }}
-                        <input type="submit" value="Sign Up!" class="btn btn-lg btn-primary btn-block text-uppercase"/>
-                    </form>
-
-                </div>
-            </div>
-        </div>
-        <div class="col-md-2"></div>
-    </div>
-{% endblock content %}
