--- users/admin.py
@@ -1,14 +1,0 @@
-from django.contrib import admin
-from django.contrib.auth.admin import UserAdmin
-
-from .forms import CustomUserCreationForm, CustomUserChangeForm
-from .models import CustomUser
-# Register your models here.
-
-class CustomUserAdmin(UserAdmin):
-    add_form = CustomUserCreationForm
-    form = CustomUserChangeForm
-    list_display = ['username', 'first_name', 'last_name', 'memory', 'country']
-    model = CustomUser
-
-admin.site.register(CustomUser, CustomUserAdmin)
