--- users/views.py
@@ -1,12 +1,0 @@
-from django.shortcuts import render
-
-from django.urls import reverse_lazy
-from django.views.generic import CreateView
-
-from .forms import CustomUserCreationForm
-# Create your views here.
-
-class SignUp(CreateView):
-    form_class = CustomUserCreationForm
-    success_url = reverse_lazy('login')
-    template_name = 'signup.html'
